Added report module to sum event durations
The system stores raw events. To make the reporting clearer we want to
convert that to durations spent on an activity. If you spend time on
an activity multiple times it merges it into one number of the total
time.

The plan is to update the report UI to output this but first I want
to refactor how the models, controllers and views are constructed for
a cleaner setup.

I may later add reporting of the amount of idle time. Not sure yet.