Refactor to use MVC design
I was kind of doing this already. The models were at the root level,
the controllers were named after the window they were creating and the
views (react components) where named after the window with the suffix
`_ui`.

But I would prefer this separate be explicit based on the directory
structure so I moved some things around.

Also I didn't like the way the
views were accessing the model directly. It created confusing
communication (such as the idle cancel). Also it might present a problem
if both the controllers and views were to access the same model since
they execute in different processes.

To resolve I made the view just responsible for two things:

* Rendering its state.
* Sending events when the user does something

It is the responsibility for the controller to load the data and hand
that to the view. I created a wrapper around setState so the controller
can set the views state.