Initial auto-complete support
Not sure this is 100% solid but it's all the time I have for today and
it's good enough for basic usage.

I originally wanted to use the HTML5 datalist element. Unfortunately
this is not supported by Electron[1]. It appears it is tied in with the
browser's auto-complete, saved forms, etc so it's not trivial to
support.

That issue pointed me to using awesomplete as a polyfill. Unfortuantly
it's not really a polyfill. It fails in the following two way:

* It doesn't monitor the `datalist` and refresh the options if I update
  the options in `datalist`.
* The list is presented within the HTML content meaning that if the
  window is small the options will be clipped (unlike a select list for
  example).

For now I am still using it as I like the fact that it doesn't depend
on external libraries and is relatively simple.

To accommodate the window size issue I just made the window bigger. I
may experiment with using a transparent window to make this better
later.

I'm unsure if I like the fact that the suggestion list state is being
maintained outside of react. The library sort of assumes it maintains
it's own state. Also other native controls such as select lists maintain
their own states so why can't this. But it still feels a bit wrong that
it's not going through the React component state.

1. https://github.com/electron/electron/issues/360