Auto-insert a idle method on quit
If I shut down the app without a final idle event there is not record
of how long my last activity was. This corrects this by trying to
always cap a day.

There is still possibly an issue of a crash and no boot later in that
day causing the same issue but this is good enough for now.