Fix auto-complete not updating state with chosen suggestion
The autocomplete tool was updating the input with the selected value but
was not sending a `change` event so this update to the input field was
not getting communicated to the React component state.

This corrects the problem. Ideally we should update awesomplete to send
a proper event but I'm not convinced I will be sticking with awesomplete
long term so hacking around it for now.