Start on documentation
This is not really complete yet. In fact the documentation refers to
features that don't yet exist. But I find in writing I have to take
a stab at it then come back later to edit. So this is just a first
draft. Once I knock through a few more features I'll come back and
edit it up.

Also included a document to make it clear this is public domain
software using the common name `UNLICENSE` for the file indicating this.