Further optimize auto-complete
I'm changing my mind about autocomplete event and onchange event being
the same. To make this clear I have rebound to the onkeyup event in
case awesomplete ever does trigger onchange.

For keyup we are getting new suggestions and setting the state. For
the autocomplete we are setting the state and submitting the value.

This auto-submit means I don't need to both accept the auto-complete
and submit. Both actions are done by a single enter key. Also the
fact that it highlights the first means the most likely auto-complete
is already selected so I don't need to first hit down.

The one case is when my value is a subset of an auto-complete value. I
think this will be rare. If it does happen you can hit ESC to remove
the auto-complete options and then hit enter to accept your subset.