Auto-remove entries more than 10 days old
There are two purposes here. The first is that we don't want people to
use this as a reporting system. You should transfer your time into a
time reporting system if that is needed. This avoids us going down the
road of providing lots of time reporting which keeps input the
optimizing focus.

This also has an input UX benefit. Projects form a year ago that are
no longer relevant won't show up in the auto-complete. We only see
recent projects. Obviously we could add a date filter but since we don't
want a history this keeps things simple.