Style report
I wanted to use less which electron-compile supports but not if the
CSS is added dynamically. This means plain CSS. Due to this and wanting
to limit the scope of my design fiddling I have kept with a pretty
basic styling but it is certainly MUCH better than the unstyled before.

Also now indicates the total activity time which is nice so I can see
how my day is going.