Escape input for suggestion search
The user input was being converted directly into a regexp. This means
if the user typed something like `[` they got an error since the
constructed regexp is now invalid.

To fix I am escaping their input. This uses the rejected polyfill[1]
for the Regexp.escape function. While I dislike using a rejected
standard there isn't yet an accepted one and the rejected one seems
quite reasonable to me.

This polyfill is affecting the global runtime but I don't think that
is such an issue in this small of an app.

This closes #8.

1. https://github.com/benjamingr/RegExp.escape/issues/37